refactor: ♻️ refactor project to get test coverage
