+++ .github/workflows/ci.yml
@@ -9,12 +9,13 @@ jobs:
     strategy:
       fail-fast: false
       matrix: 
-        python: ['3.8', '3.9']
+        python: ['3.9']
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python }}
-      - run: python test_math.py
+      - run: pip install coverage 
+      - run: python -m coverage run -m unittest
       - run: python math.py
